feat(geometry): Add geometry, revise state, model, tests.
