+++ honeybee_schema/model.py
@@ -11,60 +11,7 @@ from .energy.properties import ShadeEnergyPropertiesAbridged, \
     FaceEnergyPropertiesAbridged, RoomEnergyPropertiesAbridged, \
     ModelEnergyProperties
 
-
-class Plane(NoExtraBaseModel):
-
-    type: constr(regex='^Plane$') = 'Plane'
-
-    n: List[float] = Field(
-        ...,
-        description="Plane normal as 3 (x, y, z) values.",
-        min_items=3,
-        max_items=3
-    )
-
-    o: List[float] = Field(
-        ...,
-        description="Plane origin as 3 (x, y, z) values",
-        min_items=3,
-        max_items=3
-    )
-
-    x: List[float] = Field(
-        default=None,
-        description="Plane x-axis as 3 (x, y, z) values. If None, it is autocalculated.",
-        min_items=3,
-        max_items=3
-    )
-
-
-class Face3D(NoExtraBaseModel):
-    """A single planar face in 3D space."""
-
-    type: constr(regex='^Face3D$') = 'Face3D'
-
-    boundary: List[conlist(float, min_items=3, max_items=3)] = Field(
-        ...,
-        min_items=3,
-        description='A list of points representing the outer boundary vertices of '
-            'the face. The list should include at least 3 points and each point '
-            'should be a list of 3 (x, y, z) values.'
-    )
-
-    holes: List[conlist(conlist(float, min_items=3, max_items=3), min_items=3)] = Field(
-        default=None,
-        description='Optional list of lists with one list for each hole in the face.'
-            'Each hole should be a list of at least 3 points and each point a list '
-            'of 3 (x, y, z) values. If None, it will be assumed that there are no '
-            'holes in the face.'
-    )
-
-    plane: Plane = Field(
-        default=None,
-        description='Optional Plane indicating the plane in which the face exists.'
-            'If None, the plane will usually be derived from the boundary points.'
-    )
-
+from .geometry import Plane, Face3D
 
 class ShadePropertiesAbridged(BaseModel):
