+++ honeybee_schema/radiance/state.py
@@ -1,17 +1,28 @@
 """State Schema"""
 
 from pydantic import Field, constr
-from typing import List, Union
-from .._base import NoExtraBaseModel
+from typing import List
+from .._base import NoExtraBaseModel, IDdBaseModel
 
-from ..model import Face3D
-from .properties import _PropertiesBaseAbridged
+from ..geometry import Face3D
 
 
-class StateShade(_PropertiesBaseAbridged):
-    """StateShade represents shading for States."""
+class StateGeometryAbridged(IDdBaseModel):
+    """StateGeometryAbridged represents shading for States."""
 
-    type: constr(regex='^StateShade$') = 'StateShade'
+    type: constr(regex='^StateGeometryAbridged$') = 'StateGeometryAbridged'
+
+    modifier: str = Field(
+        default=None,
+        description='A string for a Honeybee Radiance Modifier (default: None).'
+        )
+
+    modifier_blk: str = Field(
+        default=None,
+        description='A string for a Honeybee Radiance Modifier to be used '
+                    'in direct solar simulations and in isolation studies (assessing'
+                    'the contribution of individual objects) (default: None).'
+        )
 
     geometry: Face3D = Field(
         ...,
@@ -38,9 +49,9 @@ class RadianceShadeStateAbridged(_RadianceStateBaseAbridged):
 
     type: constr(regex='^RadianceShadeStateAbridged$') = 'RadianceShadeStateAbridged'
 
-    shades: List[Union[StateShade]] = Field(
+    shades: List[StateGeometryAbridged] = Field(
         default=None,
-        description='A list of StateShade objects (default: None).'
+        description='A list of StateGeometryAbridged objects (default: None).'
     )
 
 
@@ -50,9 +61,9 @@ class RadianceSubFaceStateAbridged(_RadianceStateBaseAbridged):
 
     type: constr(regex='^RadianceSubFaceStateAbridged$') = 'RadianceSubFaceStateAbridged'
 
-    shades: List[Union[StateShade]] = Field(
+    shades: List[StateGeometryAbridged] = Field(
         default=None,
-        description='A list of StateShade objects (default: None).'
+        description='A list of StateGeometryAbridged objects (default: None).'
     )
 
     vmtx_geometry: Face3D = Field(
@@ -60,7 +71,7 @@ class RadianceSubFaceStateAbridged(_RadianceStateBaseAbridged):
         description='A Face3D for the view matrix geometry (default: None).'
     )
 
-    vmtx_geometry: Face3D = Field(
+    dmtx_geometry: Face3D = Field(
         default=None,
         description='A Face3D for the daylight matrix geometry (default: None).'
     )
